per_BirthYear smallint(4) unsigned

+++ src/mysql/upgrade/5.7.0.sql
@@ -1,1 +1,1 @@
-alter table person_per modify per_BirthYear int null;
+alter table person_per modify per_BirthYear smallint(4) unsigned null;
